Fix the forward declarations for the 'heev()' and 'heevd()' LAPACK wrapper functions

+++ blaze/math/lapack/clapack/heev.h
@@ -86,7 +86,7 @@ inline void heev( char jobz, char uplo, int n, complex<float>* A, int lda,
                   float* w, complex<float>* work, int lwork, float* rwork, int* info );
 
 inline void heev( char jobz, char uplo, int n, complex<double>* A, int lda,
-                  double* w, complex<double>* work, int lwork, float* rwork, int* info );
+                  double* w, complex<double>* work, int lwork, double* rwork, int* info );
 //@}
 //*************************************************************************************************
